requirements.txt
coverage==7.5.2
pytest==7.4.3
pytest-cov==5.0.0
pytest-mock==3.14.0
PyGithub==1.59.0
